+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v 0.3.0
+- Removed `streaming-iterator` 🤦‍
+- Fixed algorithm R (`unweighted::r`) and added back to unweighted API
+- Separated `unweighted` into feature (which is included by default)
+- Added rudimentary test suite (currently panics and prints results into console, so human can verify that results look random)
+- Added advice to use algorithm L by default in README.
+- Added "Quickstart" section to README
+
 # v 0.2.3
 - Fixed `streaming_iterator` support not working due to out-of-scope `use` statement.
 
@@ -7,5 +15,5 @@
 # v 0.2.1   
 - Removed unneccesary dependence on `compare`
 
-# 0.2.0
+# v 0.2.0
 - Added `streaming-iterator` support